Relocate get_credential_proposal_messages
